WebAppYG/requirements.txt
# Core Django
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1

# GeoDjango dependencies
psycopg2-binary==2.9.7
Pillow==10.0.0

# Database
django-extensions==3.2.3

# Task queue
celery==5.3.1
redis==4.6.0
django-celery-beat==2.5.0
django-celery-results==2.5.1

# API documentation
drf-spectacular==0.26.4

# Development
python-decouple==3.8
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0

# Data processing
pandas==2.0.3
geopandas==0.13.2
numpy==1.24.3
openpyxl==3.1.2

# API filtering
django-filter==23.5

# Utilities
python-dateutil==2.8.2
pytz==2023.3
